upp.src: Fixed various warnings (GCC 8.2)
git-svn-id: svn://ultimatepp.org/upp/trunk@12866 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/CodeEditor/DiffSyntax.cpp
@@ -79,7 +79,7 @@ bool DiffSyntax::IsPattern(const wchar *current, const wchar *end, String patter
 	
 	int i = 0;
 	while((current < end) && (i < pattern.GetCount())) {
-		if(ToLower(*current) !=  ToLower(pattern[i])) {
+		if((int)ToLower(*current) != ToLower(pattern[i])) {
 			containing = false;
 			break;
 		}
